Appease NetBSD Lint (again)
lzpack.c(1081): warning: suggest cast from 'int' to 'long' on op '+' to avoid overflow [324]

+++ lzpack.c
@@ -1078,7 +1078,7 @@ min_gap (const unsigned char *pl, long pl_len, long outlen, int litcnt,
          long pl_dst_top)
 {
   long src_base = pl_dst_top + 1 - pl_len;
-  long dst_base = TPA + litcnt;
+  long dst_base = (long)TPA + litcnt;
   long pi = 0;
   int bc = 0;
   unsigned bv = 0;
